src/core/audio_processor.py
# File: src/core/audio_processor.py
"""
Handles audio file operations including export and quality preservation.

Filepath: src/core/audio_processor.py
"""
from typing import List
import wave

class AudioProcessor:
    def __init__(self, output_dir: str):
        self.output_dir = output_dir
        
    def export_audio_clips(self, clips: List[Dict]):
        """Export audio clips while preserving quality"""
        # TODO: Implement audio export logic
        pass
    
    def validate_audio_files(self) -> bool:
        """Validate exported audio files"""
        # TODO: Implement validation logic
        pass