tests/test_audio_processor.py
# File: tests/test_audio_processor.py
"""
Tests for audio processing functionality.

Filepath: tests/test_audio_processor.py
"""
import pytest
from src.core.audio_processor import AudioProcessor

def test_audio_export(tmp_path):
    """Test exporting audio clips"""
    processor = AudioProcessor(tmp_path)
    # TODO: Add assertions for audio export

def test_audio_quality_preservation(tmp_path):
    """Test preservation of audio quality during export"""
    processor = AudioProcessor(tmp_path)
    # TODO: Add assertions for quality checks

def test_batch_processing(tmp_path):
    """Test processing multiple audio files"""
    processor = AudioProcessor(tmp_path)
    # TODO: Add assertions for batch processing